modified:   00-archives/high-badge.svg
Re-created this high-badge again. I knew there was a lot of messy history behind that file. So i restarted from scratch. Different technique. That's now made of the "Lm" line on top of the green circle, and then applying the path difference filter (bottom minus top). Thus, it becomes a circle with a cut out from the "Lm". And then there is a duplicate of this origin filled circle, behind this green path. But that was resulting in a very thin halo, I had to make that white background circle 0.5px smaller to compensate. Looked at 600% it is not perfect, but almost. This troubles comes from Inkscape, which isn't always of absolute precision...