ring-name.svg: Name re-created using a special edition Ubuntu font with variable weight. This font was created upon a download from the official https://design.ubuntu.com/font. This font was a temporary creation to get the same font shape and weight as the original typography, ± 1 pixel. Now this font is not there anymore, as the "linuxmint" text got converted to path.
Those 9 letters are now converted to path and combined into a single path. And this complex path doesn't need any stroke width to match the logo's weight or path thickness, as opposed to the previous version. So, it can be resized or recoloured without having to care about separate settings for fills and stroke. There isn't any stroke being used. And the nodes are very clean, as opposed to some other methods tested before.

To sum up, it is now made pure, simple, and strong.